backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Karn Wong <karn@karnwong.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.2"
fastapi = "^0.112.0"
psycopg = "^3.2.1"
psycopg-binary = "^3.2.1"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.21"
uvicorn = "^0.30.5"
requests = "^2.32.3"

[tool.poetry.group.tests.dependencies]
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
